iter9 - refactored filestroage
